requirements.txt
# Core
streamlit==1.39.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36

# ML
scikit-learn==1.5.2
joblib==1.4.2

# AWS / SageMaker
boto3==1.35.60
sagemaker==2.232.2

# GCP / Genai
google-genai>=0.3.0

# Data fetching
requests==2.32.3
beautifulsoup4==4.12.3
sec-edgar-downloader==5.0.3

# Utilities
tabulate==0.9.0
